requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
bcrypt==4.0.1
cached-property==1.5.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.104.1
FastAPI-SQLAlchemy==0.2.1
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
inflection==0.5.1
makefun==1.15.2
Mako==1.3.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
passlib==1.7.4
psycopg2==2.9.9
pycparser==2.21
pydantic==1.10.13
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlalchemy-orm==1.2.10
starlette==0.27.0
typing-inspect==0.9.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
